Mci is broken - gpio set in sim.py never received in store.c poll thread, means that read complete signal never sent and mutex in project.write never unlocks. All others work though
